partly rollback of switching to atomic variables
Switching to atomic operations where it was possible
in commit 35de2b904b4a3937b05423dc39717808b74a8801 was not too wise.
Some of the changes were either errorneous or lead to efficiency reducing.
This commit fix this problematic area.

This commit also
 * fully switch function.c to atomic operations
 * fix missprint in description of stat_workaround_depth in
   smnbetfs options list
 * describe smb_timeout option in smbnetfs options list

+++ src/function.h
@@ -5,7 +5,7 @@
 
 extern	struct fuse_operations	smb_oper;
 
-int function_set_free_space_size(size_t blocks_count);
+int function_set_free_space_size(int blocks_count);
 int function_set_quiet_flag(int flag);
 int function_set_dollar_share_visibility(int flag);
 int function_set_hidden_hosts_visibility(int flag);
